requirements.txt
streamlit==1.31.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.13.0
matplotlib==3.7.1
requests==2.31.0
h5py==3.9.0
